Refactor with PluginConfig
