+++ plugin/src/main/kotlin/de/fayard/BuildSrcVersionsTask.kt
@@ -1,34 +1,20 @@
 package de.fayard
 
-import com.squareup.moshi.JsonAdapter
-import com.squareup.moshi.Moshi
-import okio.buffer
-import okio.source
 import org.gradle.api.DefaultTask
 import org.gradle.api.Project
+import org.gradle.api.tasks.Input
 import org.gradle.api.tasks.TaskAction
 import org.gradle.kotlin.dsl.getByType
-import java.io.File
 
 @Suppress("UnstableApiUsage")
 open class BuildSrcVersionsTask : DefaultTask() {
 
-    companion object {
-        val moshiAdapter: JsonAdapter<DependencyGraph> by lazy {
-            Moshi.Builder().build().adapter(DependencyGraph::class.java)
-        }
-
-        fun readGraphFromJsonFile(jsonInput: File): DependencyGraph {
-            return moshiAdapter.fromJson(jsonInput.source().buffer())!!
-        }
-
-    }
-
-    var jsonInputPath = "build/dependencyUpdates/report.json"
+    @Input
+    var jsonInputPath = PluginConfig.BENMANES_REPORT_PATH
 
     @TaskAction
     fun taskAction() {
-        val extension = project.extensions.getByType<BuildSrcVersionsExtension>()
+        val extension : BuildSrcVersionsExtension = project.extensions.getByType()
         println("Configuration: $extension")
 
         val jsonInput = project.file(jsonInputPath)
@@ -39,8 +25,8 @@ open class BuildSrcVersionsTask : DefaultTask() {
         checkIfFilesExistInitially(project)
 
         val initializationMap = mapOf(
-            OutputFile.BUILD to INITIAL_BUILD_GRADLE_KTS,
-            OutputFile.GIT_IGNORE to INITIAL_GITIGNORE
+            OutputFile.BUILD to PluginConfig.INITIAL_BUILD_GRADLE_KTS,
+            OutputFile.GIT_IGNORE to PluginConfig.INITIAL_GITIGNORE
         )
 
         for ((outputFile, initialContent) in initializationMap) {
@@ -50,11 +36,11 @@ open class BuildSrcVersionsTask : DefaultTask() {
             }
         }
 
-        val dependencyGraph = readGraphFromJsonFile(jsonInput)
+        val dependencyGraph = PluginConfig.readGraphFromJsonFile(jsonInput)
 
         val useFdqnByDefault = extension.useFdqnFor.map(::escapeName)
 
-        val dependencies: List<Dependency> = parseGraph(dependencyGraph, useFdqnByDefault + MEANING_LESS_NAMES)
+        val dependencies: List<Dependency> = parseGraph(dependencyGraph, useFdqnByDefault + PluginConfig.MEANING_LESS_NAMES)
 
         val kotlinPoetry: KotlinPoetry = kotlinpoet(dependencies, dependencyGraph.gradle)
 
@@ -72,31 +58,5 @@ open class BuildSrcVersionsTask : DefaultTask() {
     }
 
 
-}
-
-internal enum class OutputFile(val path: String, var existed: Boolean = false, val alternativePath: String? = null) {
-    OUTPUTDIR("buildSrc/src/main/kotlin"),
-    BUILD("buildSrc/build.gradle.kts", alternativePath = "buildSrc/build.gradle"),
-    GIT_IGNORE("buildSrc/.gitignore"),
-    LIBS("buildSrc/src/main/kotlin/Libs.kt"),
-    VERSIONS("buildSrc/src/main/kotlin/Versions.kt");
-
-    fun fileExists(project: Project) = when {
-        project.file(path).exists() -> true
-        alternativePath != null -> project.file(alternativePath).exists()
-        else -> false
-    }
-
-    fun logFileWasModified() {
-        val ANSI_RESET = "\u001B[0m"
-        val ANSI_GREEN = "\u001B[32m"
-
-        val status = if (existed) {
-            "        modified:   "
-        } else {
-            "        new file:   "
-        }
-        println("$ANSI_GREEN$status$path$ANSI_RESET")
-    }
 }
