+++ plugin/src/main/kotlin/de/fayard/KotlinPoetry.kt
@@ -8,54 +8,6 @@ internal val LibsClassName = "Libs"
 internal val VersionsClassName = "Versions"
 
 
-/**
- * We don't want to use meaningless generic libs like Libs.core
- *
- * Found many inspiration for bad libs here https://developer.android.com/jetpack/androidx/migrate
- * **/
-val MEANING_LESS_NAMES: List<String> = listOf(
-    "common", "core", "core-testing", "testing", "runtime", "extensions",
-    "compiler", "migration", "db", "rules", "runner", "monitor", "loader",
-    "media", "print", "io", "media", "collection", "gradle", "android"
-)
-
-val INITIAL_GITIGNORE = """
-.gradle/
-build/
-"""
-
-val GRADLE_KDOC = """
-See issue 19: How to update Gradle itself?
-https://github.com/jmfayard/buildSrcVersions/issues/19
-"""
-
-val KDOC_LIBS = """
-    Generated by https://github.com/jmfayard/buildSrcVersions
-
-    Update this file with
-      `$ ./gradlew buildSrcVersions`
-    """.trimIndent()
-
-val KDOC_VERSIONS = """
-    Generated by https://github.com/jmfayard/buildSrcVersions
-
-    Find which updates are available by running
-        `$ ./gradlew buildSrcVersions`
-    This will only update the comments.
-
-    YOU are responsible for updating manually the dependency version.
-    """.trimIndent()
-
-
-const val INITIAL_BUILD_GRADLE_KTS = """
-plugins {
-    `kotlin-dsl`
-}
-repositories {
-    mavenCentral()
-}
-        """
-
 
 @Suppress("LocalVariableName")
 fun kotlinpoet(versions: List<Dependency>, gradleConfig: GradleConfig): KotlinPoetry {
@@ -69,18 +21,18 @@ fun kotlinpoet(versions: List<Dependency>, gradleConfig: GradleConfig): KotlinPo
         .map(Dependency::generateLibsProperty)
 
     val gradleProperties: List<PropertySpec> = listOf(
-        constStringProperty("gradleLatestVersion", gradleConfig.current.version, CodeBlock.of(GRADLE_KDOC)),
+        constStringProperty("gradleLatestVersion", gradleConfig.current.version, CodeBlock.of(PluginConfig.GRADLE_KDOC)),
         constStringProperty("gradleCurrentVersion", gradleConfig.running.version)
     )
 
     val Versions: TypeSpec = TypeSpec.objectBuilder("Versions")
-        .addKdoc(KDOC_VERSIONS)
+        .addKdoc(PluginConfig.KDOC_VERSIONS)
         .addProperties(versionsProperties + gradleProperties)
         .build()
 
 
     val Libs = TypeSpec.objectBuilder("Libs")
-        .addKdoc(KDOC_LIBS)
+        .addKdoc(PluginConfig.KDOC_LIBS)
         .addProperties(libsProperties)
         .build()
 
@@ -163,7 +115,7 @@ fun Dependency.generateLibsProperty(): PropertySpec {
 }
 
 
-fun BuildSrcVersionsTask.Companion.parseGraph(
+fun parseGraph(
     graph: DependencyGraph,
     useFdqnByDefault: List<String>
 ): List<Dependency> {
