+++ plugin/src/main/kotlin/de/fayard/OutputFile.kt
@@ -1,0 +1,29 @@
+package de.fayard
+
+import org.gradle.api.Project
+
+internal enum class OutputFile(val path: String, var existed: Boolean = false, val alternativePath: String? = null) {
+    OUTPUTDIR("buildSrc/src/main/kotlin"),
+    BUILD("buildSrc/build.gradle.kts", alternativePath = "buildSrc/build.gradle"),
+    GIT_IGNORE("buildSrc/.gitignore"),
+    LIBS("buildSrc/src/main/kotlin/Libs.kt"),
+    VERSIONS("buildSrc/src/main/kotlin/Versions.kt");
+
+    fun fileExists(project: Project) = when {
+        project.file(path).exists() -> true
+        alternativePath != null -> project.file(alternativePath).exists()
+        else -> false
+    }
+
+    fun logFileWasModified() {
+        val ANSI_RESET = "\u001B[0m"
+        val ANSI_GREEN = "\u001B[32m"
+
+        val status = if (existed) {
+            "        modified:   "
+        } else {
+            "        new file:   "
+        }
+        println("$ANSI_GREEN$status$path$ANSI_RESET")
+    }
+}
